+++ packages/BoDialog/src/main.vue
@@ -6,202 +6,41 @@
     @close="closeDialog"
     :close-on-click-modal="false"
   >
-    <el-form
-      v-if="visible"
-      v-bind="form"
-      ref="dialogForm"
-      :label-width="form.labelWidth"
-      :label-position="labelPosition"
-      @submit.native.prevent
-    >
-      <el-form-item
-        v-for="(item, i) in formItems"
-        :key="`${i}item`"
-        v-show="item.showOn ? item.showOn === form.type : true"
-        v-bind="item"
+    <transition name="dialog-fade">
+      <el-form
+        v-if="visible"
+        v-bind="form"
+        ref="dialogForm"
+        :label-width="form.labelWidth"
+        :label-position="labelPosition"
+        @submit.native.prevent
       >
-        <el-input
-          v-if="item.itemType === 'input' || item.itemType === undefined"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          clearable
-          :type="item.type || 'text'"
-          :placeholder="item.placeholder"
-          :disabled="form.type === item.disabledOn"
-          @blur="form.model[item.prop] = $event.target.value.trim()"
-        />
-        <bo-currency-input
-          v-else-if="item.itemType === 'currency'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          clearable
-          :placeholder="item.placeholder"
-          :disabled="form.type === item.disabledOn"
-        ></bo-currency-input>
-        <div v-else-if="item.itemType === 'text'">
-          {{ form.model[item.prop] }}
-        </div>
-        <el-select
-          v-else-if="item.itemType === 'select'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :placeholder="item.placeholder"
-          :multiple="item.multiple"
-          :disabled="form.type === item.disabledOn"
-          clearable
+        <el-form-item
+          v-for="(item, i) in formItems"
+          :key="`${i}item`"
+          v-show="item.showOn ? item.showOn === form.type : true"
+          v-bind="item"
         >
-          <el-option
-            v-for="(option, optionIndex) in item.options"
-            :key="optionIndex + '_local'"
-            :value="option.value"
-            :label="option.label"
-          />
-        </el-select>
-        <bo-select
-          v-else-if="item.itemType === 'multSelect'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :list="item.options"
-          :disabled="form.type === item.disabledOn"
-        >
-        </bo-select>
-        <bo-mult-lang
-          v-else-if="item.itemType === 'multLang'"
-          v-model="form.model[item.prop]"
-          :langs="item.langs"
-          :form="item.form"
-          :formItems="item.formItems"
-          :disabled="form.type === item.disabledOn"
-        >
-        </bo-mult-lang>
-        <el-date-picker
-          v-else-if="item.itemType === 'date'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          type="date"
-          :placeholder="item.placeholder"
-          clearable
-          :picker-options="item.pickerOptions || {}"
-          :disabled="form.type === item.disabledOn"
-        />
-        <el-date-picker
-          v-else-if="item.itemType === 'datetime'"
-          v-model="form.model[item.prop]"
-          type="datetime"
-          :placeholder="item.placeholder"
-          :format="item.format || 'yyyy-MM-dd HH:mm:ss'"
-          :value-format="item.valueFormat || 'yyyy-MM-dd HH:mm:ss'"
-          clearable
-          :picker-options="item.pickerOptions || {}"
-          :disabled="form.type === item.disabledOn"
-        />
-        <el-date-picker
-          v-else-if="item.itemType === 'monthrange'"
-          v-model="form.model[item.prop]"
-          type="monthrange"
-          :clearable="false"
-          :value-format="item.valueFormat || 'yyyyMM'"
-          :picker-options="item.pickerOptions || monthPickerOptions"
-          :disabled="form.type === item.disabledOn"
-          style="width: 280px"
-        />
-        <el-date-picker
-          v-else-if="item.itemType === 'daterange'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          type="daterange"
-          value-format="yyyyMMdd"
-          :clearable="false"
-          :placeholder="item.placeholder"
-          :picker-options="item.pickerOptions || pickerOptions"
-          :disabled="form.type === item.disabledOn"
-          style="width: 280px"
-        />
-        <el-switch
-          v-else-if="item.itemType === 'switch'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-        ></el-switch>
-        <el-checkbox-group
-          v-else-if="item.itemType === 'checkbox'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :disabled="form.type === item.disabledOn"
-        >
-          <el-checkbox
-            v-for="(checkbox, checkboxIndex) in item.options"
-            :key="checkboxIndex"
-            v-bind="checkbox"
-          ></el-checkbox>
-        </el-checkbox-group>
-        <el-radio-group
-          v-else-if="item.itemType === 'radio'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :disabled="form.type === item.disabledOn"
-        >
-          <el-radio
-            v-for="(radio, radioIndex) in item.options"
-            :key="radioIndex"
-            :label="radio.value"
-            v-bind="radio"
-            >{{ radio.label }}</el-radio
-          >
-        </el-radio-group>
-        <el-input
-          v-else-if="item.itemType === 'textarea'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :show-word-limit="item.showWordLimit"
-          :size="item.size"
-          :rows="item.rows"
-          :autosize="item.autosize"
-          :disabled="form.type === item.disabledOn"
-          type="textarea"
-          @blur="form.model[item.prop] = $event.target.value.trim()"
-        ></el-input>
-        <bo-emoji
-          v-else-if="item.itemType === 'emoji'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :show-word-limit="item.showWordLimit"
-          :size="item.size"
-          :rows="item.rows"
-          :autosize="item.autosize"
-          :disabled="form.type === item.disabledOn"
-        ></bo-emoji>
-        <bo-image-upload
-          v-else-if="item.itemType === 'imageUpload'"
-          v-model="form.model[item.prop]"
-          @input="onChange(item)"
-          :disabled="form.type === item.disabledOn"
-        ></bo-image-upload>
-        <!-- @slot 可從外部傳入 slot 給 itemType: slot 使用 -->
-        <slot
-          v-else-if="item.itemType === 'slot'"
-          :name="item.slotName"
-          :model="form.model"
-        ></slot>
-        <div class="tips">
-          {{ item.tips }}
-        </div>
-      </el-form-item>
-    </el-form>
+          <bo-form-item
+            v-model="form.model"
+            :item="item"
+            :form="form"
+            @input="onChange(item)"
+          ></bo-form-item>
+        </el-form-item>
+      </el-form>
+    </transition>
     <div slot="footer" v-if="showFooter">
       <el-button @click="closeDialog">Cancel</el-button>
-      <el-button type="primary" @click="confirm" :loading="loading"
-        >{{ confirmTitle }}</el-button
-      >
+      <el-button type="primary" @click="confirm" :loading="loading">{{
+        confirmTitle
+      }}</el-button>
     </div>
   </el-dialog>
 </template>
 
 <script>
-import BoSelect from "../../BoSelect";
-import BoCurrencyInput from "../../BoCurrencyInput";
-import BoMultLang from "../../BoMultLang";
-import BoEmoji from "../../BoEmoji"
-import BoImageUpload from "../../BoImageUpload"
+import BoFormItem from "../../BoFormItem/index";
 import { pickerOptionsData } from "../../config/picker-options";
 import { setFormMsg, setFormItemsMsg } from "../../utils/formDefaultMessage";
 /**
@@ -210,11 +49,7 @@ import { setFormMsg, setFormItemsMsg } from "../../utils/formDefaultMessage";
 export default {
   name: "BoDialog",
   components: {
-    BoSelect,
-    BoCurrencyInput,
-    BoMultLang,
-    BoEmoji,
-    BoImageUpload,
+    BoFormItem,
   },
   computed: {
     title() {
@@ -243,7 +78,9 @@ export default {
           const { formItems } = this.$props;
 
           Object.keys(this.form.model).forEach((key) => {
-            this.form.model[key] = pickerOptionsData(formItems).params[key] || getDefault(this.form.model[key]);
+            this.form.model[key] =
+              pickerOptionsData(formItems).params[key] ||
+              getDefault(this.form.model[key]);
           });
         }
       },
@@ -259,7 +96,7 @@ export default {
       handler(formItems) {
         setFormItemsMsg(formItems, "Please check the configuration.");
       },
-    }
+    },
   },
   props: {
     /**
@@ -301,12 +138,12 @@ export default {
      */
     confirmTitle: {
       type: String,
-      default: 'Confirm',
+      default: "Confirm",
     },
     showFooter: {
       type: Boolean,
-      default: true
-    }
+      default: true,
+    },
   },
   data() {
     const { formItems } = this.$props;
@@ -373,9 +210,3 @@ export default {
   },
 };
 </script>
-<style lang="scss" scoped>
-.tips {
-  color: #ccc;
-  font-size: 12px;
-}
-</style>