+++ packages/BoFormItem/index.js
@@ -1,0 +1,7 @@
+import BoFormItem from './src/main';
+
+BoFormItem.install = function(Vue) {
+  Vue.component(BoFormItem.name, BoFormItem)
+};
+
+export default BoFormItem;