+++ packages/BoRangeInput/index.js
@@ -1,0 +1,7 @@
+import BoRangeInput from './src/main';
+
+BoRangeInput.install = function(Vue) {
+  Vue.component(BoRangeInput.name, BoRangeInput)
+};
+
+export default BoRangeInput;